# ReachInbox-Email-Scheduler-1/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * ReachInbox outbound email scheduling API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Error {
  error: string;
}

export type CampaignStatus = typeof CampaignStatus[keyof typeof CampaignStatus];


export const CampaignStatus = {
  draft: 'draft',
  running: 'running',
  paused: 'paused',
  completed: 'completed',
} as const;

export type ProspectStatus = typeof ProspectStatus[keyof typeof ProspectStatus];


export const ProspectStatus = {
  active: 'active',
  replied: 'replied',
  bounced: 'bounced',
  unsubscribed: 'unsubscribed',
} as const;

export type ActivityType = typeof ActivityType[keyof typeof ActivityType];


export const ActivityType = {
  sent: 'sent',
  opened: 'opened',
  replied: 'replied',
  bounced: 'bounced',
} as const;

export interface Campaign {
  id: number;
  name: string;
  audience: string;
  status: CampaignStatus;
  stepCount: number;
  sentCount: number;
  openRate: number;
  replyRate: number;
  /** @nullable */
  nextSendAt: string | null;
  createdAt: string;
}

export interface CampaignInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  audience: string;
}

export interface CampaignUpdate {
  /** @minLength 1 */
  name?: string;
  /** @minLength 1 */
  audience?: string;
}

export interface CampaignStatusUpdate {
  status: CampaignStatus;
}

export interface SequenceStep {
  id: number;
  campaignId: number;
  position: number;
  subject: string;
  body: string;
  /** @minimum 0 */
  delayDays: number;
  sentCount: number;
  openRate: number;
  replyRate: number;
}

export interface CampaignDetail {
  campaign: Campaign;
  steps: SequenceStep[];
}

export interface SequenceStepInput {
  /** @minLength 1 */
  subject: string;
  /** @minLength 1 */
  body: string;
  /** @minimum 0 */
  delayDays: number;
}

export interface SequenceStepUpdate {
  /** @minLength 1 */
  subject?: string;
  /** @minLength 1 */
  body?: string;
  /** @minimum 0 */
  delayDays?: number;
}

export interface Prospect {
  id: number;
  name: string;
  email: string;
  company: string;
  role: string;
  status: ProspectStatus;
  /** @nullable */
  campaignId: number | null;
  /** @nullable */
  lastContactedAt: string | null;
  createdAt: string;
}

export interface ProspectInput {
  /** @minLength 1 */
  name: string;
  email: string;
  /** @minLength 1 */
  company: string;
  /** @minLength 1 */
  role: string;
  /** @nullable */
  campaignId?: number | null;
}

export interface ProspectUpdate {
  /** @minLength 1 */
  name?: string;
  email?: string;
  /** @minLength 1 */
  company?: string;
  /** @minLength 1 */
  role?: string;
  status?: ProspectStatus;
  /** @nullable */
  campaignId?: number | null;
}

export interface Activity {
  id: number;
  type: ActivityType;
  message: string;
  prospectName: string;
  campaignName: string;
  occurredAt: string;
}

export interface TrendPoint {
  label: string;
  sent: number;
  opened: number;
  replied: number;
}

export interface DashboardSummary {
  totalProspects: number;
  activeCampaigns: number;
  emailsSent: number;
  openRate: number;
  replyRate: number;
  scheduledToday: number;
  trend: TrendPoint[];
}

/**
 * Resource not found
 */
export type NotFoundResponse = Error;

export type LimitParameter = number;

export type ListActivitiesParams = {
/**
 * @minimum 1
 * @maximum 100
 */
limit?: LimitParameter;
};

export type ListCampaignsParams = {
status?: CampaignStatus;
search?: string;
};

export type ListProspectsParams = {
status?: ProspectStatus;
campaignId?: number;
search?: string;
};