# ReachInbox-Email-Scheduler-1/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ReachInbox outbound email scheduling API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: dashboard
    description: Workspace overview
  - name: campaigns
    description: Email campaigns and sequences
  - name: prospects
    description: Prospect records
  - name: activity
    description: Recent sending activity
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [dashboard]
      summary: Get dashboard summary
      responses:
        "200":
          description: Dashboard metrics and recent trend
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"
  /activities:
    get:
      operationId: listActivities
      tags: [activity]
      summary: List recent activity
      parameters:
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Recent activity items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Activity"
  /campaigns:
    get:
      operationId: listCampaigns
      tags: [campaigns]
      summary: List campaigns
      parameters:
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/CampaignStatus"
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Campaign list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Campaign"
    post:
      operationId: createCampaign
      tags: [campaigns]
      summary: Create a campaign
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CampaignInput"
      responses:
        "201":
          description: Created campaign
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
  /campaigns/{id}:
    get:
      operationId: getCampaign
      tags: [campaigns]
      summary: Get a campaign with its sequence
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Campaign detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CampaignDetail"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      operationId: updateCampaign
      tags: [campaigns]
      summary: Update a campaign
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CampaignUpdate"
      responses:
        "200":
          description: Updated campaign
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteCampaign
      tags: [campaigns]
      summary: Delete a campaign
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "204":
          description: Campaign deleted
        "404":
          $ref: "#/components/responses/NotFound"
  /campaigns/{id}/status:
    patch:
      operationId: updateCampaignStatus
      tags: [campaigns]
      summary: Change a campaign status
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CampaignStatusUpdate"
      responses:
        "200":
          description: Campaign with updated status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
        "404":
          $ref: "#/components/responses/NotFound"
  /campaigns/{id}/steps:
    get:
      operationId: listCampaignSteps
      tags: [campaigns]
      summary: List steps for a campaign
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "200":
          description: Sequence steps
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SequenceStep"
        "404":
          $ref: "#/components/responses/NotFound"
    post:
      operationId: createCampaignStep
      tags: [campaigns]
      summary: Add a step to a campaign
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SequenceStepInput"
      responses:
        "201":
          description: Created sequence step
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SequenceStep"
        "404":
          $ref: "#/components/responses/NotFound"
  /campaigns/{id}/steps/{stepId}:
    patch:
      operationId: updateCampaignStep
      tags: [campaigns]
      summary: Update a sequence step
      parameters:
        - $ref: "#/components/parameters/Id"
        - $ref: "#/components/parameters/StepId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SequenceStepUpdate"
      responses:
        "200":
          description: Updated sequence step
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SequenceStep"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteCampaignStep
      tags: [campaigns]
      summary: Delete a sequence step
      parameters:
        - $ref: "#/components/parameters/Id"
        - $ref: "#/components/parameters/StepId"
      responses:
        "204":
          description: Sequence step deleted
        "404":
          $ref: "#/components/responses/NotFound"
  /prospects:
    get:
      operationId: listProspects
      tags: [prospects]
      summary: List prospects
      parameters:
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/ProspectStatus"
        - name: campaignId
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Prospect list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Prospect"
    post:
      operationId: createProspect
      tags: [prospects]
      summary: Add a prospect
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProspectInput"
      responses:
        "201":
          description: Created prospect
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Prospect"
  /prospects/{id}:
    patch:
      operationId: updateProspect
      tags: [prospects]
      summary: Update a prospect
      parameters:
        - $ref: "#/components/parameters/Id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProspectUpdate"
      responses:
        "200":
          description: Updated prospect
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Prospect"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteProspect
      tags: [prospects]
      summary: Delete a prospect
      parameters:
        - $ref: "#/components/parameters/Id"
      responses:
        "204":
          description: Prospect deleted
        "404":
          $ref: "#/components/responses/NotFound"
components:
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
    StepId:
      name: stepId
      in: path
      required: true
      schema:
        type: integer
        minimum: 1
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    Error:
      type: object
      properties:
        error:
          type: string
      required: [error]
    CampaignStatus:
      type: string
      enum: [draft, running, paused, completed]
    ProspectStatus:
      type: string
      enum: [active, replied, bounced, unsubscribed]
    ActivityType:
      type: string
      enum: [sent, opened, replied, bounced]
    Campaign:
      type: object
      required:
        - id
        - name
        - audience
        - status
        - stepCount
        - sentCount
        - openRate
        - replyRate
        - nextSendAt
        - createdAt
      properties:
        id: { type: integer }
        name: { type: string }
        audience: { type: string }
        status: { $ref: "#/components/schemas/CampaignStatus" }
        stepCount: { type: integer }
        sentCount: { type: integer }
        openRate: { type: number }
        replyRate: { type: number }
        nextSendAt: { type: ["string", "null"], format: date-time }
        createdAt: { type: string, format: date-time }
    CampaignInput:
      type: object
      required: [name, audience]
      properties:
        name: { type: string, minLength: 1 }
        audience: { type: string, minLength: 1 }
    CampaignUpdate:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        audience: { type: string, minLength: 1 }
    CampaignStatusUpdate:
      type: object
      required: [status]
      properties:
        status: { $ref: "#/components/schemas/CampaignStatus" }
    CampaignDetail:
      type: object
      required: [campaign, steps]
      properties:
        campaign: { $ref: "#/components/schemas/Campaign" }
        steps:
          type: array
          items: { $ref: "#/components/schemas/SequenceStep" }
    SequenceStep:
      type: object
      required:
        - id
        - campaignId
        - position
        - subject
        - body
        - delayDays
        - sentCount
        - openRate
        - replyRate
      properties:
        id: { type: integer }
        campaignId: { type: integer }
        position: { type: integer }
        subject: { type: string }
        body: { type: string }
        delayDays: { type: integer, minimum: 0 }
        sentCount: { type: integer }
        openRate: { type: number }
        replyRate: { type: number }
    SequenceStepInput:
      type: object
      required: [subject, body, delayDays]
      properties:
        subject: { type: string, minLength: 1 }
        body: { type: string, minLength: 1 }
        delayDays: { type: integer, minimum: 0 }
    SequenceStepUpdate:
      type: object
      properties:
        subject: { type: string, minLength: 1 }
        body: { type: string, minLength: 1 }
        delayDays: { type: integer, minimum: 0 }
    Prospect:
      type: object
      required:
        - id
        - name
        - email
        - company
        - role
        - status
        - campaignId
        - lastContactedAt
        - createdAt
      properties:
        id: { type: integer }
        name: { type: string }
        email: { type: string, format: email }
        company: { type: string }
        role: { type: string }
        status: { $ref: "#/components/schemas/ProspectStatus" }
        campaignId: { type: ["integer", "null"] }
        lastContactedAt: { type: ["string", "null"], format: date-time }
        createdAt: { type: string, format: date-time }
    ProspectInput:
      type: object
      required: [name, email, company, role]
      properties:
        name: { type: string, minLength: 1 }
        email: { type: string, format: email }
        company: { type: string, minLength: 1 }
        role: { type: string, minLength: 1 }
        campaignId: { type: ["integer", "null"] }
    ProspectUpdate:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        email: { type: string, format: email }
        company: { type: string, minLength: 1 }
        role: { type: string, minLength: 1 }
        status: { $ref: "#/components/schemas/ProspectStatus" }
        campaignId: { type: ["integer", "null"] }
    Activity:
      type: object
      required: [id, type, message, prospectName, campaignName, occurredAt]
      properties:
        id: { type: integer }
        type: { $ref: "#/components/schemas/ActivityType" }
        message: { type: string }
        prospectName: { type: string }
        campaignName: { type: string }
        occurredAt: { type: string, format: date-time }
    TrendPoint:
      type: object
      required: [label, sent, opened, replied]
      properties:
        label: { type: string }
        sent: { type: integer }
        opened: { type: integer }
        replied: { type: integer }
    DashboardSummary:
      type: object
      required:
        - totalProspects
        - activeCampaigns
        - emailsSent
        - openRate
        - replyRate
        - scheduledToday
        - trend
      properties:
        totalProspects: { type: integer }
        activeCampaigns: { type: integer }
        emailsSent: { type: integer }
        openRate: { type: number }
        replyRate: { type: number }
        scheduledToday: { type: integer }
        trend:
          type: array
          items: { $ref: "#/components/schemas/TrendPoint" }