# ReachInbox-Email-Scheduler-1/lib/api-zod/src/generated/types/activity.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * ReachInbox outbound email scheduling API
 * OpenAPI spec version: 0.1.0
 */
import type { ActivityType } from './activityType';

export interface Activity {
  id: number;
  type: ActivityType;
  message: string;
  prospectName: string;
  campaignName: string;
  occurredAt: Date;
}