# ReachInbox-Email-Scheduler-1/lib/api-zod/src/generated/types/activityType.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * ReachInbox outbound email scheduling API
 * OpenAPI spec version: 0.1.0
 */

export type ActivityType = typeof ActivityType[keyof typeof ActivityType];


export const ActivityType = {
  sent: 'sent',
  opened: 'opened',
  replied: 'replied',
  bounced: 'bounced',
} as const;