# ReachInbox-Email-Scheduler-1/lib/api-zod/src/generated/types/campaignDetail.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * ReachInbox outbound email scheduling API
 * OpenAPI spec version: 0.1.0
 */
import type { Campaign } from './campaign';
import type { SequenceStep } from './sequenceStep';

export interface CampaignDetail {
  campaign: Campaign;
  steps: SequenceStep[];
}