# ReachInbox-Email-Scheduler-1/lib/api-zod/src/generated/types/campaignInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * ReachInbox outbound email scheduling API
 * OpenAPI spec version: 0.1.0
 */

export interface CampaignInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  audience: string;
}