# ReachInbox-Email-Scheduler-1/lib/api-zod/src/generated/types/campaignStatus.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * ReachInbox outbound email scheduling API
 * OpenAPI spec version: 0.1.0
 */

export type CampaignStatus = typeof CampaignStatus[keyof typeof CampaignStatus];


export const CampaignStatus = {
  draft: 'draft',
  running: 'running',
  paused: 'paused',
  completed: 'completed',
} as const;