# ReachInbox-Email-Scheduler-1/lib/api-zod/src/generated/types/campaignStatusUpdate.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * ReachInbox outbound email scheduling API
 * OpenAPI spec version: 0.1.0
 */
import type { CampaignStatus } from './campaignStatus';

export interface CampaignStatusUpdate {
  status: CampaignStatus;
}