# ReachInbox-Email-Scheduler-1/lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * ReachInbox outbound email scheduling API
 * OpenAPI spec version: 0.1.0
 */
import type { TrendPoint } from './trendPoint';

export interface DashboardSummary {
  totalProspects: number;
  activeCampaigns: number;
  emailsSent: number;
  openRate: number;
  replyRate: number;
  scheduledToday: number;
  trend: TrendPoint[];
}