# ReachInbox-Email-Scheduler-1/lib/api-zod/src/generated/types/prospect.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * ReachInbox outbound email scheduling API
 * OpenAPI spec version: 0.1.0
 */
import type { ProspectStatus } from './prospectStatus';

export interface Prospect {
  id: number;
  name: string;
  email: string;
  company: string;
  role: string;
  status: ProspectStatus;
  /** @nullable */
  campaignId: number | null;
  /** @nullable */
  lastContactedAt: Date | null;
  createdAt: Date;
}