# ReachInbox-Email-Scheduler-1/lib/api-zod/src/generated/types/prospectUpdate.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * ReachInbox outbound email scheduling API
 * OpenAPI spec version: 0.1.0
 */
import type { ProspectStatus } from './prospectStatus';

export interface ProspectUpdate {
  /** @minLength 1 */
  name?: string;
  email?: string;
  /** @minLength 1 */
  company?: string;
  /** @minLength 1 */
  role?: string;
  status?: ProspectStatus;
  /** @nullable */
  campaignId?: number | null;
}