# ReachInbox-Email-Scheduler-1/lib/api-zod/src/generated/types/trendPoint.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * ReachInbox outbound email scheduling API
 * OpenAPI spec version: 0.1.0
 */

export interface TrendPoint {
  label: string;
  sent: number;
  opened: number;
  replied: number;
}